requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sentence-transformers
scikit-learn
pdfminer.six
python-docx
jinja2
